HW-0000 Travel agency service
- created sln file
