+++ TravelAgencyService/TravelAgencyService/Models/AgencyModel.cs
@@ -4,7 +4,7 @@ namespace TravelAgencyService.Models;
 
 public class AgencyModel
 {
-    private readonly List<Destination> _destinations =
+    private readonly List<Destination?> _destinations =
     [
         new Destination("Spain", 5),
         new Destination("Ukraine", 3),
@@ -12,16 +12,13 @@ public class AgencyModel
         new Destination("Norway", 2),
         new Destination("New Zealand", 3)
     ];
-
-    private ExceptionMgmt _exceptionMgmt;
-
     public void BookTicket(string destinationName)
     {
         var destination = FindDestination(destinationName);
 
             if (destination == null)
             {
-                throw new ExceptionMgmt.InvalidDestinationException("Wrong destination. Please, choose the right one.");
+                throw new InvalidDestinationException("Wrong destination. Please, choose the right one.");
             }
 
         destination.BookTicket();
@@ -30,21 +27,16 @@ public class AgencyModel
     public bool CancelBooking(string destinationName)
     {
         var destination = FindDestination(destinationName);
-
-            if (destination != null)
-            {
-                return destination.CancelBooking();
-            }
-        return false;
+        return destination != null && destination.CancelBooking();
     }
 
-    public List<Destination> GetAllBookings()
+    public List<Destination?> GetAllBookings()
     {
-        return _destinations.Where(d => d.CurrentBookings > 0).ToList();
+        return _destinations.Where(d => d is { CurrentBookings: > 0 }).ToList();
     }
 
-    private Destination FindDestination(string destinationName)
+    private Destination? FindDestination(string destinationName)
     {
-        return _destinations.FirstOrDefault(d => d.Name.Equals(destinationName, StringComparison.OrdinalIgnoreCase));
+        return _destinations.FirstOrDefault(d => d != null && d.Name.Equals(destinationName, StringComparison.OrdinalIgnoreCase));
     }
 }
